requirements.txt
opencv-python
tk
Pillow
psutil
GPUtil
ultralytics
torch
numpy